# services/faux-schoolday/docs/api-spec.yaml
openapi: 3.0.3
info:
  title: Faux Schoolday API
  description: |
    High-fidelity mock implementation of Schoolday platform APIs for teacher evaluation system integration.

    This mock service provides comprehensive simulation of:
    - OAuth 2.0 authentication flows
    - OneRoster v1.1 standard for educational data exchange
    - Discovery API for district and school information
    - Academy API for professional development
    - Webhook system for real-time event notifications

    All APIs are designed to match Schoolday's actual behavior for seamless local development and testing.
  version: 1.0.0
  contact:
    name: Teacher Evaluation Platform Team
    email: dev@teacher-eval.local
  license:
    name: MIT License
servers:
  - url: http://localhost:3006/api
    description: Local development server
  - url: https://faux-schoolday.teacher-eval.local/api
    description: Containerized environment
security:
  - BearerAuth: []
paths:
...
  /academy/courses:
    get:
      tags: [Academy]
      summary: List professional development courses
      parameters:
        - name: category
          in: query
          schema:
            type: string
          example: Instructional Design
        - name: level
          in: query
          schema:
            type: string
            enum: [Beginner, Intermediate, Advanced]
        - name: format
          in: query
          schema:
            type: string
            enum: [Online, Hybrid, In-Person]
        - name: search
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: List of courses
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseListResponse'
  /academy/recommendations:
    get:
      tags: [Academy]
      summary: Get course recommendations
      description: AI-powered course recommendations for teachers
      parameters:
        - name: teacher_id
          in: query
          required: true
          schema:
            type: string
        - name: subject_area
          in: query
          schema:
            type: string
        - name: skill_level
          in: query
          schema:
            type: string
        - name: career_goals
          in: query
          schema:
            type: string
        - name: time_availability
          in: query
          schema:
            type: integer
          description: Hours per week
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: Course recommendations
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseRecommendationsResponse'

  # Webhook API Endpoints
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Course:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        category:
          type: string
        level:
          type: string
          enum: [Beginner, Intermediate, Advanced]
        format:
          type: string
          enum: [Online, Hybrid, In-Person]
        duration_hours:
          type: integer
        price:
          type: number
        rating:
          type: string
        description:
          type: string
        tags:
          type: array
          items:
            type: string
        next_start_date:
          type: string
          format: date
    CourseListResponse:
      type: object
      properties:
        courses:
          type: array
          items:
            $ref: '#/components/schemas/Course'
        pagination:
          $ref: '#/components/schemas/Pagination'
        categories:
          type: array
          items:
            type: string
    CourseRecommendation:
      allOf:
        - $ref: '#/components/schemas/Course'
        - type: object
          properties:
            recommendation_score:
              type: number
            recommendation_reasons:
              type: array
              items:
                type: string
            estimated_impact:
              type: object
              properties:
                teaching_effectiveness:
                  type: string
                student_engagement:
                  type: string
                career_advancement:
                  type: string
    CourseRecommendationsResponse:
      type: object
      properties:
        recommendations:
          type: array
          items:
            $ref: '#/components/schemas/CourseRecommendation'
        recommendation_metadata:
          type: object
          properties:
            teacher_id:
              type: string
            algorithm_version:
              type: string
            generated_at:
              type: string
              format: date-time

    # Webhook Schemas
...
    Pagination:
      type: object
      properties:
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
        hasMore:
          type: boolean
...